refactor: LottoSystem 책임 분리 리팩토링
